refactor: 폴더 이동 및 코드 정리
